+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - NELO-1: Initialize project. Switch from `npm` to `yarn`
 - NELO-2: Implement ZSA and create first Server Action `getPokemonList`
+- NELO-3: Implement Logic of fetch the details of each Pokemon in the map